fixed header img innerheight
The window's load event was nested into a function that was loaded only
once and then cached.  subsequent load events wouldn't fire.  the header
image's innerHeight was calculated on load and so the image wouldn't
load unless a full refresh without the browsers cache was performed.  I
moved the windows load event to the loader.js file where it would fire
on every refresh.